pyproject.toml
[project]
name = "image-manip"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pillow (>=11.1.0,<12.0.0)",
    "pydantic>=2.11.7",
]

[[project.authors]]
name = "Natoandro"

[dependency-groups]
dev = [
    "pytest>=8.4.1",
]
